chapter_13/sniper/requirements.txt
ebaysdk
click
plyer
pyobjus
beautifulsoup4
pyyaml
python-cron
asyncio_pool
httpx
httpx[http2]
